fixed Windows build; various cosmetic changes

+++ include/nxconfig.h
@@ -158,11 +158,11 @@ public:
    ConfigMergeStrategy getMergeStrategy() const { return m_mergeStrategy; }
 
 	bool loadXmlConfig(const TCHAR *file, const char *topLevelTag = NULL, bool merge = true);
-	bool loadXmlConfigFromMemory(const char *xml, int xmlSize, const TCHAR *name = NULL, const char *topLevelTag = NULL, bool merge = true);
+	bool loadXmlConfigFromMemory(const char *xml, size_t xmlSize, const TCHAR *name = NULL, const char *topLevelTag = NULL, bool merge = true);
 	bool loadIniConfig(const TCHAR *file, const TCHAR *defaultIniSection, bool ignoreErrors = true);
    bool loadIniConfigFromMemory(const char *content, size_t length, const TCHAR *fileName, const TCHAR *defaultIniSection, bool ignoreErrors = true);
 	bool loadConfig(const TCHAR *file, const TCHAR *defaultIniSection, const char *topLevelTag = NULL, bool ignoreErrors = true, bool merge = true);
-	bool loadConfigFromMemory(const char *xml, int xmlSize, const TCHAR *defaultIniSection, const char *topLevelTag, bool ignoreErrors, bool merge);
+	bool loadConfigFromMemory(const char *xml, size_t xmlSize, const TCHAR *defaultIniSection, const char *topLevelTag, bool ignoreErrors, bool merge);
 
 	bool loadConfigDirectory(const TCHAR *path, const TCHAR *defaultIniSection, const char *topLevelTag = NULL, bool ignoreErrors = true, bool merge = true);
